desktop/requirements.txt
PyQt6==6.10.2
PyMuPDF==1.27.2
python-docx==1.2.0
